requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-ckeditor==6.0.0
django-currentuser==0.5.1
django-debug-toolbar==3.1.1
django-js-asset==1.2.2
django-modeltranslation==0.16
django-mptt==0.11.0
django-recaptcha3==0.4.0
django-redis==4.12.1
flower==0.9.5
humanize==3.1.0
idna==2.10
kombu==4.6.11
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pytils==0.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tornado==6.0.4
urllib3==1.25.11
vine==1.3.0
wcwidth==0.2.5
wheel==0.35.1
